Initial commit
This code is for plotting the expression of genes like VRS1 (two- vs. six-rowed barley), INT-C (Intermedium-C aka Teosinte Branched 1), and VRN3 (Vernalization 3 aka Flowering Locus Type 1). It could stand to be cleaned up a little bit, but important things like the gene names, input files, and core code are there so it is functional.